feat(account) email
refactor: 기존 Email API Endpoint EmailController로 병합
