+++ src/test/java/com/tune_fun/v1/account/adapter/input/rest/EmailControllerIT.java
@@ -5,6 +5,8 @@ import com.tune_fun.v1.base.ControllerBaseTest;
 import com.tune_fun.v1.common.config.Uris;
 import com.tune_fun.v1.common.response.MessageCode;
 import com.tune_fun.v1.dummy.DummyService;
+import com.tune_fun.v1.otp.adapter.output.persistence.OtpType;
+import com.tune_fun.v1.otp.domain.value.CurrentDecryptedOtp;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
@@ -16,12 +18,14 @@ import org.springframework.transaction.annotation.Transactional;
 import static com.epages.restdocs.apispec.ResourceDocumentation.resource;
 import static com.epages.restdocs.apispec.ResourceSnippetParameters.builder;
 import static com.tune_fun.v1.base.doc.RestDocsConfig.constraint;
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.queryParameters;
 
-class CheckEmailDuplicateControllerIT extends ControllerBaseTest {
+class EmailControllerIT extends ControllerBaseTest {
 
     @Autowired
     private DummyService dummyService;
@@ -84,6 +88,41 @@ class CheckEmailDuplicateControllerIT extends ControllerBaseTest {
                                 )
                         )
                 );
+    }
+
+    @Transactional
+    @Test
+    @Order(3)
+    @DisplayName("이메일 인증 여부 확인, 성공")
+    void checkEmailVerifiedSuccess() throws Exception {
+        dummyService.initAccount();
+        dummyService.forgotPasswordOtp();
+
+        CurrentDecryptedOtp forgotPasswordOtp = dummyService.getForgotPasswordOtp();
+        dummyService.verifyOtp(OtpType.FORGOT_PASSWORD, forgotPasswordOtp.token());
+
+        dummyService.login(dummyService.getDefaultAccount());
+        String accessToken = dummyService.getDefaultAccessToken();
+
+        mockMvc.perform(
+                        get(Uris.CHECK_EMAIL_VERIFIED)
+                                .header(AUTHORIZATION, bearerToken(accessToken))
+                )
+                .andExpectAll(baseAssertion(MessageCode.SUCCESS_EMAIL_VERIFIED))
+                .andDo(
+                        restDocs.document(
+                                requestHeaders(authorizationHeader),
+                                responseFields(baseResponseFields),
+                                resource(
+                                        builder().
+                                                description("이메일 인증 여부 확인").
+                                                requestHeaders(authorizationHeader).
+                                                responseFields(baseResponseFields)
+                                                .build()
+                                )
+                        )
+                );
+
     }
 
 }
